Fixed data scope of arrays passed into functions
Previously, array data was copied into function parameters, so any
changes to the array within the function were not reflected on the
variable passed as the argument. The correct behaviour in C is to pass a
reference/pointer to the argument instead, which is now what happens.
Possibly not the neatest implementation, but seems to work...