Added support for distinct 'float' and 'double' types
Previously, there was a single "floating point' type, that was just
treated as a double. Adding distinct types allows for more flexibility
in their usage, as well as proper precision for various benchmarks.

Type promotion in expressions is performed as if FLT_EVAL_METHOD=0 in
'float.h', i.e. if all values are float then float will be used
internally for calculations, otherwise double will be used.