Fixed parsing of #define, #include and pragmas after if/for/do/while
Previously these could be used as the statement executed by a
conditional or loop, which differs from the behaviour of standard C
(which would have removed these statements in the preprocessor). Now
the parser will keep executing until it hits the end of the first proper
statement after the start of a conditional or loop.